+++ Changelog.md
@@ -1,6 +1,7 @@
 ## v1.2.0
 - Screen stays on when Fullscreen player is visible
 - New feature, music of the day: A notification that everyday around 7 AM suggests a random music
+- New setting: Background cover blur radius
 - UI improvements
 
 ## v1.1.1